+++ src/components/Fund/ListBox.tsx
@@ -13,17 +13,18 @@ const people = [
   { name: 'Hellen Schmidt' },
 ]
 
-const ListBox = () => {
-    const [selected, setSelected] = useState(people[0])
+const ListBox = (props: any) => {
+    const { funds } = props;
+    const [selected, setSelected] = useState(funds[0])
   
     return (
       <div className="relative">
         <Listbox value={selected} onChange={setSelected}>
           <div className="mt-1">
-            <Listbox.Button className="w-3/5 rounded-2xl bg-black px-4 py-2.5 text-sm font-medium text-white shadow-sm hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 focus:ring-offset-orange-100">
+            <Listbox.Button className="w-3/5 rounded-2xl bg-black px-4 py-2.5 text-sm font-medium text-white shadow-sm hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-offset-2 focus:ring-offset-orange-100">
                 <div className='flex justify-between'>
-                    <div className='space-x-2 items-center'>
-                        {/* <img src={presentFund?.logo} alt="" /> */}
+                    <div className='flex items-center space-x-2'>
+                        <img src={selected?.logo} alt="" />
                         <p className='text-2xl font-bold'>{selected?.name}</p>
                     </div>
                     <img className='block' src={dropDownArrow} alt="" />
@@ -35,29 +36,30 @@ const ListBox = () => {
               leaveFrom="opacity-100"
               leaveTo="opacity-0"
             >
-              <Listbox.Options className="absolute mt-1 max-h-60 w-3/5 overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                {people.map((person, personIdx) => (
+              <Listbox.Options className="absolute mt-1 max-h-60 w-3/5 overflow-auto rounded-md bg-gray-800 py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
+                {funds.map((fund: any) => (
                   <Listbox.Option
-                    key={personIdx}
+                    key={fund?.id}
                     className={({ active }) =>
-                      `relative cursor-default select-none py-2 pl-10 pr-4 ${
-                        active ? 'bg-amber-100 text-amber-900' : 'text-gray-900'
+                      `relative cursor-default select-none py-2 pl-3 pr-4 ${
+                        active ? 'bg-amber-100 text-amber-900' : 'text-white'
                       }`
                     }
-                    value={person}
+                    value={fund}
                   >
                     {({ selected }) => (
                       <>
-                        <span
-                          className={`block truncate ${
-                            selected ? 'font-medium' : 'font-normal'
+                        <div
+                          className={`truncate flex items-center text-lg space-x-3 ${
+                            selected ? 'font-bold' : 'font-normal'
                           }`}
                         >
-                          {person.name}
-                        </span>
+                          <img src={fund?.logo} alt="" />  
+                          <p>{fund?.name}</p>
+                        </div>
                         {selected ? (
                           <span className="absolute inset-y-0 left-0 flex items-center pl-3 text-amber-600">
-                            {/* <CheckIcon className="h-5 w-5" aria-hidden="true" /> */}
+                            
                           </span>
                         ) : null}
                       </>
